Modern ES imports, Update package.json
